Major defence system update
Added multiple pieces into the defence system triggers, so it now shows
a couple more perception defences and added a Summoner section with all
demons as well as class specific defences. Also added the mass and cloak
defences into the mini-console on the bottom as well as demons. When
demons are available, they show up green and red when they're unleashed.

Also some circle additions and changes.