Small but important tweak and update
Basic circle updates and also a very significant addition of macros so
that you don't have to always change them by hand when transferring the
system between Windows and Mac. The bindings are still the same CTRL+key
but in Mac CTRL is interpreted as META, so those have been added.